Pull in `host`.
